% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_senate_results_by.R
\name{get_senate_results_by}
\alias{get_senate_results_by}
\title{Obtain senate results}
\usage{
get_senate_results_by(year = c("2016", "2013", "2010", "2007", "2004"),
  by = c("division", "polling_place"), state = c("NSW", "VIC", "QLD", "SA",
  "WA", "TAS", "ACT", "NT"))
}
\arguments{
\item{year}{The year of the election for which results are desired.}

\item{by}{Which table of results should be fetched? \code{division} for results aggregated up to division; \code{polling_place} to obtain the results at each polling place.}

\item{state}{Which seats are results requested? Setting \code{by = 'division'} while selecting a state will ignore the state argument, with a warning.}
}
\value{
A \code{data.table} of the results.
}
\description{
Obtain senate results
}
\note{
Connects to AEC website.
}

