% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_region.R
\name{plot_region}
\alias{plot_region}
\title{Plot region of Australia}
\usage{
plot_region(city, coord_lines = c("none", "latlon", "lat", "lon"),
  inset_scale = 15)
}
\arguments{
\item{city}{Which city to plot}

\item{coord_lines}{Plot grid lines at 0.1 intervals.}

\item{inset_scale}{Scale relative to Australia.}
}
\value{
A list of three elements, \code{vwidth} the width of the inset on the device, \code{vheight} the height of the inset on the device, and \code{p} the plot.
}
\description{
Plot region of Australia
}
\note{
This function should not be called directly. It is intended to be used to plot insets on a map generated through \code{\link{draw_electoral_divisions_map}}.
}

