% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-polling_places_2016.R
\docType{data}
\name{polling_places_2016}
\alias{polling_places_2016}
\title{Polling place information for 2016}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 8328 rows and 15 columns.}
\usage{
polling_places_2016
}
\description{
Polling place information for 2016
}
\keyword{datasets}

