% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_city_plot.R
\name{print_city_plot}
\alias{print_city_plot}
\title{Print inset of city}
\usage{
print_city_plot(.polygonal.data, city, xpos, ypos, vfactor, ...)
}
\arguments{
\item{.polygonal.data}{Polygonal data of the entire Australian map.}

\item{city}{A city code.}

\item{xpos}{The x-positions (in \code{npc}) that the inset is the be drawn.}

\item{ypos}{The y-position.}

\item{vfactor}{By how much should the city inset be scaled \strong{relative to other cities}.}

\item{...}{Passed to \code{\link{plot_region}}.}
}
\description{
Print inset of city
}
\note{
This function must not be called directly.
}

