% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_electoral_divisions_map.R
\name{draw_electoral_divisions_map}
\alias{draw_electoral_divisions_map}
\title{Draw an electoral map, with insets.}
\usage{
draw_electoral_divisions_map(.polygonal.data, year = c("2016", "2013", "2010",
  "2007", "2004"), scale_fill_manual_args = NULL,
  polygon_outline_colour = "black", city_inset_scale = 12,
  legend.title = NULL)
}
\arguments{
\item{.polygonal.data}{Data including the vertices of all polygons, including a \code{fill} variable to colour the polygons.}

\item{year}{What elections' boundaries should be used?}

\item{scale_fill_manual_args}{A list of arguments to pass to \code{scale_fill_manual}.}

\item{polygon_outline_colour}{The outline of polygons.}

\item{city_inset_scale}{By how much should the cities' insets be scaled by relative to the Australia plot?}

\item{legend.title}{What should the legend title (for \code{fill}) be?}
}
\value{
A printed ggplot object.
}
\description{
Draw an electoral map, with insets.
}

