% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_region.R
\name{plot_region}
\alias{plot_region}
\title{Plot region of Australia}
\usage{
plot_region(.polygonal.data, city = c("PER", "ADL", "MEL", "SYD", "BNE"),
  coord_lines = c("none", "latlon", "lat", "lon"),
  scale_fill_manual_args = NULL, polygon_outline_colour = "black",
  city_inset_scale = 15, title = TRUE)
}
\arguments{
\item{.polygonal.data}{Polygonal data (of Australia). See \link{draw_electoral_divisions_map}.}

\item{city}{Which city to plot}

\item{coord_lines}{Plot grid lines at 0.1 intervals.}

\item{scale_fill_manual_args}{A list of arguments to pass to \code{scale_fill_manual}.}

\item{polygon_outline_colour}{The outline of polygons.}

\item{city_inset_scale}{Scale relative to Australia.}
}
\value{
A list of three elements, \code{vwidth} the width of the inset on the device, \code{vheight} the height of the inset on the device, and \code{p} the plot.
}
\description{
Plot region of Australia
}
\note{
This function should not be called directly. It is intended to be used to plot insets on a map generated through \code{\link{draw_electoral_divisions_map}}.
}

